+++ app.py
@@ -42,11 +42,14 @@ from urllib.request import urlopen, Request
 import nltk
 from nltk.corpus import stopwords
 
+from backend.google import (
+    google_query,
+    public_query_prefix
+)
+
 bp = Blueprint("routes", __name__, static_folder="static", template_folder="static")
 
 cosmos_db_ready = asyncio.Event()
-
-public_query_prefix = "@"
     
 def create_app():
     app = Quart(__name__)
@@ -243,220 +246,6 @@ async def init_cosmosdb_client():
 
     return cosmos_conversation_client
 
-def extract_prefix (message):
-    pattern = re.escape(public_query_prefix) + r'\w*'
-    match = re.match(pattern, message)
-    return match.group(0) if match else None
-
-def extract_time_info(message):
-    if message.startswith(public_query_prefix):
-        match = re.match(rf'{re.escape(public_query_prefix)}([dwmy]\d+)', message)
-        if match:
-            return match.group(1)
-    return None
-
-def optimize_google_query(query):
-    # Remove special characters
-    cleaned_query = re.sub(r'[^a-zA-Z0-9\s]', '', query)
-
-    # Tokenize and remove stop words
-    stop_words = set(stopwords.words('english'))
-    tokens = cleaned_query.split()
-    optimized_tokens= [word for word in tokens if word.lower() not in stop_words]
-    
-    # Join back into a string
-    optimized_query = ' '.join(optimized_tokens)
-    
-    return optimized_query
-
-async def optimize_google_query_with_ai_model(query):
-    prompt= f"""
-    You are a Google Search Assistant. You receive a prompt as an input and produce and optimized Google search query. 
-    As an output you will return only optimized query. Today's date is {datetime.now().strftime("%B %d, %Y")}
-    
-    input:
-    {query}
-    """
-    request_body = {
-        "messages": [
-            {
-                "role": "user",
-                "content": prompt
-            }
-        ]
-    }
-    try:
-        model_args= prepare_model_args(request_body, {})
-        model_args["stream"] = False
-        model_args["extra_body"] = None
-        model_args["temperature"] = 0.2
-        azure_openai_client = await init_openai_client()
-        raw_response = await azure_openai_client.chat.completions.with_raw_response.create(**model_args)
-        parsed_response = raw_response.parse()
-        return parsed_response.choices[0].message.content
-    except Exception as e:
-        logging.exception("Exception in ask_model_optimize_query")
-        return optimize_google_query(query)
-
-async def generate_google_query_parameters_with_ai_model(query):
-    prompt= f"""
-    You are an assistant that receives a prompt as an input. Because you don't have the knowledge, from the input you identify the optimal Google query you need in order to answer the prompt. 
-    You generate a json object containing the parameters that you should use for Google custom search. 
-    You should include the following parameters in the json:
-    - q: The search query
-    - num: The number of results to return (default is 5)
-    - dateRestrict: The date range for the search results (optional)
-    In addition, add a new key called "prompt" to the json object, which should contain a user prompt I can send to a GPT model so that the model can answer my query.
-    Today's date is {datetime.now().strftime("%B %d, %Y")}
-    
-    input:
-    {query}
-    """
-    request_body = {
-        "messages": [
-            {
-                "role": "user",
-                "content": prompt
-            }
-        ]
-    }
-    try:
-        model_args= prepare_model_args(request_body, {})
-        model_args["stream"] = False
-        model_args["extra_body"] = None
-        azure_openai_client = await init_openai_client()
-        raw_response = await azure_openai_client.chat.completions.with_raw_response.create(**model_args)
-        parsed_response = raw_response.parse()
-        # convert to json object 
-        json_string = re.sub(r'\n','',re.sub(r'```', '', re.sub(r'```json\n', '', parsed_response.choices[0].message.content)))
-        # string to json object
-        return json.loads(json_string)
-    except Exception as e:
-        logging.exception("Exception in ask_model_optimize_query")
-        return optimize_google_query(query)
-
-
-async def preprocess_page_content_with_ai_model(content, query):
-    prompt= f"""
-    Provide me with the information I requested, specified in Input. Use only the Sources to provide an response.
-    If the Sources is not enough to provide the requested information, please respond with an empty string.
-    Today's date is {datetime.now().strftime("%B %d, %Y")}.
-
-    Example Input:
-    What's the weather in San Francisco today?
-
-    Example Response:
-    It's 70 degrees and sunny in San Francisco today. 
-
-    Input:
-    {query}
-
-    Sources:
-    {content}
-    """
-    request_body = {
-        "messages": [
-            {
-                "role": "user",
-                "content": prompt
-            }
-        ]
-    }
-    try:
-        model_args= prepare_model_args(request_body, {})
-        model_args["stream"] = False
-        model_args["extra_body"] = None
-        model_args["temperature"] = 0
-        azure_openai_client = await init_openai_client()
-        raw_response = await azure_openai_client.chat.completions.with_raw_response.create(**model_args)
-        parsed_response = raw_response.parse()
-        return parsed_response.choices[0].message.content
-    except Exception as e:
-        logging.exception("Exception in preprocess_page_content_with_ai_model")
-        return None
-    
-def download_page(url):
-    try:
-        req=Request(url, headers={'User-Agent': 'Mozilla/5.0'})
-        html = urlopen(req, timeout=5).read()
-        soup = BeautifulSoup(UnicodeDammit.detwingle(html), features="html.parser")
-        
-        # get text
-        text = soup.get_text(strip=True)
-
-        # break into lines and remove leading and trailing space on each
-        return text[:1000000]
-        #return text
-    
-    except Exception as e:
-        logging.exception("An error occurred while fetching the page")
-        return None
-
-def fetch_google_results(query,date_restrict=None):
-    google_api_key = os.environ.get("GOOGLE_API_KEY")
-    google_cx = os.environ.get("GOOGLE_CX")
-    if not google_api_key or not google_cx:
-        raise ValueError("GOOGLE_API_KEY and GOOGLE_CX are required")
-    headers = {
-        "Content-Type": "application/json"
-    }
-    params = {
-        "key": google_api_key,
-        "cx": google_cx,
-        "q": query,
-        "num": 8,
-        "fields": "items(title,link,snippet)",
-        "lr": "lang_en",
-        "c2coff": "1"
-    }
-    if date_restrict:
-        params["dateRestrict"] = date_restrict
-    response = httpx.get("https://www.googleapis.com/customsearch/v1", headers=headers, params=params)
-    response.raise_for_status()
-
-    return response.json()
-
-def fetch_google_results_new(query_parameters):
-    google_api_key = os.environ.get("GOOGLE_API_KEY")
-    google_cx = os.environ.get("GOOGLE_CX")
-    if not google_api_key or not google_cx:
-        raise ValueError("GOOGLE_API_KEY and GOOGLE_CX are required")
-    headers = {
-        "Content-Type": "application/json"
-    }
-    query_parameters["key"] = google_api_key
-    query_parameters["cx"] = google_cx
-    query_parameters["fields"] = "items(title,link,snippet)"
-    query_parameters["lr"] = "lang_en"
-    response = httpx.get("https://www.googleapis.com/customsearch/v1", headers=headers, params=query_parameters)
-    response.raise_for_status()
-
-    return response.json()
-
-async def google_query(request_body):
-    #query_prefix = extract_prefix(request_body)
-    #date_restrict = extract_time_info(query_prefix)
-    #query = request_body.removeprefix(query_prefix).strip()
-    #optimized_query = await optimize_google_query_with_ai_model(query)
-    query = request_body.removeprefix(public_query_prefix).strip()
-    google_query_parameters = await generate_google_query_parameters_with_ai_model(query)
-    optimized_prompt = google_query_parameters.get("prompt")
-    google_query_parameters["prompt"] = None
-    google_results = fetch_google_results_new(google_query_parameters)
-    input_prompt=[]
-    if google_results.get("items"):
-        for item in google_results["items"]:
-            prompt = {
-                "url": item["link"]
-            }
-            web_page_content =  download_page(item["link"])                      
-            if web_page_content:
-                prompt["content"] = await preprocess_page_content_with_ai_model(web_page_content,optimized_prompt)
-            else:
-                prompt["content"] = item["snippet"]    
-            input_prompt.append(prompt)
-    return input_prompt
-
 def prepare_model_args(request_body, request_headers):
     request_messages = request_body.get("messages", [])
     messages = []
@@ -643,30 +432,34 @@ async def send_chat_request(request_body, request_headers):
             if message.get("role") != 'tool':
                 if message["role"] == "user" and message["content"].startswith(public_query_prefix) and i == len(messages)-1:
                     public_query = True
-                    sources = await google_query(message["content"])
-                    query = message["content"].removeprefix(extract_prefix(message["content"]))
-                        
-                    prompt =  f'''
-                        Provide me with the information I requested. Use the sources to provide an accurate response.
-                        Respond in markdown format. Cite the sources you used as a markdown link as you use them at the 
-                        end of each sentence by number of the source (example: [[1]](link.com)). Make sure the references
-                        are counted properly and are in order. Provide an accurate response and then stop. 
-                        Today's date is {datetime.now().strftime("%B %d, %Y")}.
+                    result = await google_query(message["content"])
+                    if result:
+                        query = result[1]
+                        sources = result[0]
+                            
+                        prompt =  f'''
+                            Provide me with the information I requested. Use the sources to provide an accurate response.
+                            Respond in markdown format. Cite the sources you used as a markdown link as you use them at the 
+                            end of each sentence by number of the source (example: [[1]](link.com)). Make sure the references
+                            are counted properly and are in order. Provide an accurate response and then stop. 
+                            Today's date is {datetime.now().strftime("%B %d, %Y")}.
 
-                        Example Input:
-                        What's the weather in San Francisco today?
+                            Example Input:
+                            What's the weather in San Francisco today?
 
-                        Example Response:
-                        It's 70 degrees and sunny in San Francisco today. [[1]](https://www.google.com/search?q=weather+san+francisco)
+                            Example Response:
+                            It's 70 degrees and sunny in San Francisco today. [[1]](https://www.google.com/search?q=weather+san+francisco)
 
-                        Input:
-                        {query}
+                            Input:
+                            {query}
 
-                        Sources:
-                        {sources}
+                            Sources:
+                            {sources}
 
-                    '''
-                    message["content"] = prompt
+                        '''
+                        message["content"] = prompt
+                elif message["content"].startswith(public_query_prefix):
+                    message = None
                 filtered_messages.append(message)
             
     request_body['messages'] = filtered_messages
